python-rag/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers>=2.2.2
faiss-cpu==1.7.4
whoosh==2.7.4
jieba==0.42.1
numpy==1.24.3
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv>=1.0.0